% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bs4Dash_distill.R
\name{distill_load_spinner}
\alias{distill_load_spinner}
\title{Distill load spinner}
\usage{
distill_load_spinner(ui_element, type = 4, color = NULL, ...)
}
\arguments{
\item{ui_element}{The element to apply the spinner to.}

\item{type}{The type of spinner to use, valid values are
integers between 0-8 (0 means no spinner).
Check out \url{https://daattali.com/shiny/shinycssloaders-demo}
to see the different types of spinners. You can also
use your own custom image using the image parameter.}

\item{color}{The color of the spinner in hex format.
Ignored if image is used.}

\item{...}{See \code{\link[shinycssloaders]{withSpinner}}.}
}
\description{
Distill themed \code{\link[shinycssloaders]{withSpinner}} spinner.
}
