% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_with_dashboard}
\alias{test_with_dashboard}
\title{Test distill theme with dashboard}
\usage{
test_with_dashboard()
}
\description{
Test distill theme with dashboard
}
\examples{
if (interactive()) {
   test_with_dashboard()
}
}
