% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_with_simple_app}
\alias{test_with_simple_app}
\title{Test distill theme with simple app}
\usage{
test_with_simple_app()
}
\description{
Test distill theme with simple app
}
\examples{
if (interactive()) {
   test_with_simple_app()
}
}
