% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{theme_hamilton_distill}
\alias{theme_hamilton_distill}
\title{ggplot2 Hamilton Distill Theme}
\usage{
theme_hamilton_distill()
}
\description{
Add the Hamilton Distill theme to a ggplot2 plot.
}
\examples{
ggplot2::ggplot(mtcars) +
   ggplot2::geom_point(ggplot2::aes(wt, mpg)) +
   theme_hamilton_distill()

}
