% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bs4Dash_distill.R
\name{use_bs4Dash_distill_css}
\alias{use_bs4Dash_distill_css}
\title{Use Distill CSS for bs4Dash app}
\usage{
use_bs4Dash_distill_css()
}
\description{
Use Distill CSS for bs4Dash app
}
\examples{
if (interactive()) {
  library(shiny)
  library(fresh)
  library(shiny)
  ui <- fluidPage(
    use_bs4Dash_distill_css(),
    tags$h1("Distill theme"),

    sidebarLayout(
      sidebarPanel(
        "This is the sidebar panel"
      ),
      mainPanel(
        tags$h1("First level title"),
        tags$h2("Second level title"),
        tags$h3("Third level title"),
        tags$h4("Fourth level title"),
        tags$h5("Fifth level title"),
        tags$h6("Sixth level title")
      )
    )
  )

  server <- function(input, output, session) {

  }
  shinyApp(ui, server)
}

}
